requirements.txt
RapidFuzz~=3.13.0
fastapi~=0.115.12
pillow~=11.2.1
uvicorn